[REFACTOR] Tef Shap directories and init.py
